+++ src/main/java/board/Board.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import movepolicy.rule.MoveTrace;
-import participant.Score;
-import participant.Turn;
+import core.Score;
+import core.Turn;
 import pieces.Piece;
 import pieces.Side;
 import position.Position;
