+++ src/main/java/service/PreparedGame.java
@@ -1,4 +1,6 @@
-package core;
+package service;
+
+import core.JanggiGame;
 
 public record PreparedGame(Long gameId, JanggiGame game, int initialMoveCount) {
 }
