+++ src/main/java/view/OutputView.java
@@ -1,11 +1,11 @@
 package view;
 
 import core.GameStatus;
-import core.GameSummary;
+import service.GameSummary;
 import java.util.List;
 import java.util.Map;
-import movepolicy.MoveHistory;
-import participant.Score;
+import core.MoveHistory;
+import core.Score;
 import pieces.Side;
 
 public class OutputView {
