+++ src/test/java/board/BoardTest.java
@@ -7,8 +7,8 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
-import participant.Score;
-import participant.Turn;
+import core.Score;
+import core.Turn;
 import pieces.Piece;
 import pieces.PieceType;
 import pieces.Side;
