+++ src/test/java/core/JanggiGameTest.java
@@ -9,7 +9,6 @@ import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import participant.Turn;
 import pieces.Piece;
 import pieces.PieceType;
 import pieces.Side;
