+++ src/test/java/db/repository/JdbcJanggiGameRepositoryTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 
 import board.Board;
 import core.GameStatus;
-import core.GameSummary;
+import service.GameSummary;
 import core.JanggiGame;
 import db.dao.BoardPieceDao;
 import db.dao.GameDao;
@@ -24,12 +24,12 @@ import java.sql.Statement;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import movepolicy.MoveHistory;
+import core.MoveHistory;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
-import participant.Turn;
+import core.Turn;
 import pieces.Piece;
 import pieces.PieceType;
 import pieces.Side;
